% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_names.R
\name{gene_names}
\alias{gene_names}
\title{Matrices of Matching Genes}
\usage{
gene_names(tissue_matrix_1, tissue_matrix_2)
}
\arguments{
\item{tissue_1_matrix}{A file containing gene-donor data of the first tissue}

\item{tissue_2_matrix}{A file containing gene-donor data of the second tissue}
}
\value{
Two matrices as variables in the global environment
}
\description{
Takes in two tissue data sets and attempts to match the genes within both tissues. Returns two matrices for each tissue with the matching genes.
}
