% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_pair_expression.R
\name{tissue_pair_gene_expression}
\alias{tissue_pair_gene_expression}
\title{Gene Expression Correlations Amongst Tissue Samples.}
\usage{
tissue_pair_gene_expression(tissue_1_matrix, tissue_2_matrix)
}
\arguments{
\item{tissue_1_matrix}{A file containing gene-donor data of the first tissue}

\item{tissue_2_matrix}{A file containing gene-donor data of the second tissue}
}
\value{
Biweight Mid correlations and P-values of gene comparisons across both tissue sets
}
\description{
A package that takes two gene data files and conducts a biweight Midcorrelation analysis.
The function will output two matrices containing the biweight midcorrelation correlation values and P-values.
}
