% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_filter.R
\name{var_filter}
\alias{var_filter}
\title{Variance Filter Function}
\usage{
var_filter(tissue_matrix_1, tissue_matrix_2, percentage_variance)
}
\arguments{
\item{tissue_matrix_1}{A file containing gene-donor data of the first tissue}

\item{tissue_matrix_2}{A file containing gene-donor data of the second tissue}

\item{percentage_variance}{A decimal entry representing the variance percentage for gene analysis.}
}
\description{
Functions that takes tissue matrices and variance percentage and outputs matrices of the most variable genes for each tissue set.
}
