% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{wrapper_func}
\alias{wrapper_func}
\title{geneExpCor Wrapper Function //WORKING TITLE//}
\usage{
wrapper_func(
  tissue1,
  tissue2,
  var_perc = 0.05,
  var_perc2 = 1e-04,
  var_plot = FALSE
)
}
\arguments{
\item{tissue1}{A matrix object of normalized TPM expression data, with genes as rows and sample subject identifiers as columns.}

\item{tissue2}{A matrix object of normalized TPM expression data, with genes as rows and sample subject identifiers as columns.}

\item{var_perc}{A decimal representing the percentage of highest variable genes to keep.}

\item{var_perc2}{A decimal representing the percentage of highest most significant pvalues to keep}

\item{var_plot}{Boolean value for whether you would like histogram visual of gene variances}
}
\description{
This function performs the entire tissue pair correlation analysis in a single function call.
It requires 2 gene expression matrices, a variance percentage for filtering, and the desired amount of top correlations to return.
}
