% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_names.R
\name{gene_names}
\alias{gene_names}
\title{Matrices of Matching Genes}
\usage{
gene_names(tissue_1_matrix, tissue_2_matrix)
}
\arguments{
\item{tissue_1_matrix}{A file containing gene-donor data of the first tissue}

\item{tissue_2_matrix}{A file containing gene-donor data of the second tissue}
}
\value{
Two matrices with matching donors and genes across both tissue sets. Both outputted matrices are viable for correlation analysis as inputs for the function
}
\description{
Takes in two tissue data sets and attempts to match the genes within both tissues.
This function also attempts to label the genes with the proper gene names.
Returns two matrices for each tissue with the matching genes.
}
