% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_pair_gene_expression_edit.R
\name{tissue_pair_gene_expression}
\alias{tissue_pair_gene_expression}
\title{Gene Expression Correlations Amongst Tissue Samples.}
\usage{
tissue_pair_gene_expression(tissue_1_matrix, tissue_2_matrix)
}
\arguments{
\item{tissue_1_matrix}{A file containing gene-donor data of the first tissue}

\item{tissue_2_matrix}{A file containing gene-donor data of the second tissue}
}
\value{
Biweight Mid correlations and P-values of gene comparisons across both tissue sets
}
\description{
A package that takes two gene data files and produces graphics and correlation information based on preferred method and gene variability. Methods include, 'pearson', 'Spearman', and 'Biweight MidCorrelation' where the biweight method is the default parameter. The operation and optimization of the function is still in progress
}
