% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_pair_expression.R
\name{tissue_pair_gene_expression}
\alias{tissue_pair_gene_expression}
\title{Bicor Analysis Functino //Working Title}
\usage{
tissue_pair_gene_expression(
  tissue_1_matrix,
  tissue_2_matrix,
  percentage_variance
)
}
\arguments{
\item{tissue_1_matrix}{A file containing gene-donor data of the first tissue}

\item{tissue_2_matrix}{A file containing gene-donor data of the second tissue}

\item{percentage_variance}{A decimal entry representing the variance percentage for gene analysis.}
}
\value{
A series of plots in a pdf associated with the bicor midcorrelation analysis conducted on the inputted tissue matrices
}
\description{
Takes in two tissue data sets and attempts to match the genes within both tissues.
This function also attempts to label the genes with the proper gene names.
Returns two matrices for each tissue with the matching genes.
}
